+++ crates/rtlola-frontend/src/lowering.rs
@@ -458,6 +458,8 @@ impl Mir {
             StreamAccessKind::SlidingWindow(wref) => mir::StreamAccessKind::SlidingWindow(wref),
             StreamAccessKind::Hold => mir::StreamAccessKind::Hold,
             StreamAccessKind::Offset(o) => mir::StreamAccessKind::Offset(Self::lower_offset(o)),
+            StreamAccessKind::Get => mir::StreamAccessKind::Get,
+            StreamAccessKind::Fresh => mir::StreamAccessKind::Fresh,
         }
     }