+++ crates/rtlola-hir/src/type_check/pacing_ast_climber.rs
@@ -511,7 +511,7 @@ where
                             }
                         }
                     },
-                    StreamAccessKind::Hold => {},
+                    StreamAccessKind::Hold | StreamAccessKind::Get | StreamAccessKind::Fresh => {},
                     StreamAccessKind::DiscreteWindow(_) | StreamAccessKind::SlidingWindow(_) => {
                         self.pacing_tyc
                             .impose(term_keys.exp_pacing.concretizes_explicit(Periodic(Freq::Any)))?;
@@ -520,13 +520,13 @@ where
                 };
 
                 for arg in args {
-                    let arg_key = self.expression_infer(&*arg)?;
+                    let arg_key = self.expression_infer(arg)?;
                     self.impose_more_concrete(term_keys, arg_key)?;
                 }
             },
             ExpressionKind::Default { expr, default } => {
-                let ex_key = self.expression_infer(&*expr)?;
-                let def_key = self.expression_infer(&*default)?;
+                let ex_key = self.expression_infer(expr)?;
+                let def_key = self.expression_infer(default)?;
 
                 self.impose_more_concrete(term_keys, ex_key)?;
                 self.impose_more_concrete(term_keys, def_key)?;
@@ -552,9 +552,9 @@ where
                 consequence,
                 alternative,
             } => {
-                let cond_key = self.expression_infer(&*condition)?;
-                let cons_key = self.expression_infer(&*consequence)?;
-                let alt_key = self.expression_infer(&*alternative)?;
+                let cond_key = self.expression_infer(condition)?;
+                let cons_key = self.expression_infer(consequence)?;
+                let alt_key = self.expression_infer(alternative)?;
 
                 self.impose_more_concrete(term_keys, cond_key)?;
                 self.impose_more_concrete(term_keys, cons_key)?;
@@ -568,16 +568,16 @@ where
             },
             ExpressionKind::Function(FnExprKind { args, .. }) => {
                 for arg in args {
-                    let arg_key = self.expression_infer(&*arg)?;
+                    let arg_key = self.expression_infer(arg)?;
                     self.impose_more_concrete(term_keys, arg_key)?;
                 }
             },
             ExpressionKind::TupleAccess(t, _) => {
-                let exp_key = self.expression_infer(&*t)?;
+                let exp_key = self.expression_infer(t)?;
                 self.impose_more_concrete(term_keys, exp_key)?;
             },
             ExpressionKind::Widen(hir::WidenExprKind { expr: inner, .. }) => {
-                let exp_key = self.expression_infer(&*inner)?;
+                let exp_key = self.expression_infer(inner)?;
                 self.impose_more_concrete(term_keys, exp_key)?;
             },
         };
@@ -652,6 +652,108 @@ where
             || spawn_cond != &positive_top
             || filter != &positive_top
             || close != &negative_top
+    }
+
+    fn get_or_fresh_targets(hir: &Hir<M>, expr: &Expression) -> Vec<(bool, Span, StreamReference)> {
+        match &expr.kind {
+            ExpressionKind::LoadConstant(_) => vec![],
+            ExpressionKind::ArithLog(_, children) => {
+                children
+                    .iter()
+                    .flat_map(|e| Self::get_or_fresh_targets(hir, e))
+                    .collect()
+            },
+            ExpressionKind::StreamAccess(target, kind, arguments) => {
+                let mut res: Vec<_> = arguments
+                    .iter()
+                    .flat_map(|e| Self::get_or_fresh_targets(hir, e))
+                    .collect();
+                match kind {
+                    StreamAccessKind::Get => res.push((true, expr.span.clone(), *target)),
+                    StreamAccessKind::Fresh => res.push((false, expr.span.clone(), *target)),
+                    _ => {},
+                };
+                res
+            },
+            ExpressionKind::ParameterAccess(_, _) => vec![],
+            ExpressionKind::Ite {
+                condition,
+                consequence,
+                alternative,
+            } => {
+                let mut cond = Self::get_or_fresh_targets(hir, condition);
+
+                cond.append(&mut Self::get_or_fresh_targets(hir, consequence));
+                cond.append(&mut Self::get_or_fresh_targets(hir, alternative));
+
+                cond
+            },
+            ExpressionKind::Tuple(children) => {
+                children
+                    .iter()
+                    .flat_map(|e| Self::get_or_fresh_targets(hir, e))
+                    .collect()
+            },
+            ExpressionKind::TupleAccess(target, _) => Self::get_or_fresh_targets(hir, target),
+            ExpressionKind::Function(def) => {
+                def.args
+                    .iter()
+                    .flat_map(|e| Self::get_or_fresh_targets(hir, e))
+                    .collect()
+            },
+            ExpressionKind::Widen(def) => Self::get_or_fresh_targets(hir, def.expr.as_ref()),
+            ExpressionKind::Default { expr, default } => {
+                let mut expr = Self::get_or_fresh_targets(hir, expr);
+
+                expr.append(&mut Self::get_or_fresh_targets(hir, default));
+
+                expr
+            },
+        }
+    }
+
+    fn check_get_and_fresh_access(
+        hir: &Hir<M>,
+        pacing_tt: &TypeTable<AbstractPacingType>,
+        nid_key: &HashMap<NodeId, StreamTypeKeys>,
+        expr: Option<ExprId>,
+        condition: Option<ExprId>,
+        own_pacing: &ConcretePacingType,
+    ) -> Vec<TypeError<PacingErrorKind>> {
+        expr.map(|e| Self::get_or_fresh_targets(hir, hir.expression(e)))
+            .unwrap_or_else(Vec::new)
+            .iter()
+            .chain(
+                condition
+                    .map(|e| Self::get_or_fresh_targets(hir, hir.expression(e)))
+                    .unwrap_or_else(Vec::new)
+                    .iter(),
+            )
+            .flat_map(|(is_get, span, target)| {
+                let other_pacing: &ConcretePacingType = &pacing_tt[&nid_key[&NodeId::SRef(*target)].exp_pacing];
+                if (own_pacing.is_periodic() != other_pacing.is_periodic())
+                    || (own_pacing.is_event_based() != other_pacing.is_event_based())
+                {
+                    let target_span = hir
+                        .output(*target)
+                        .map(|o| o.span.clone())
+                        .or_else(|| hir.input(*target).map(|i| i.span.clone()))
+                        .expect("Ref to be either input or output");
+                    Some(
+                        PacingErrorKind::InvalidGetOrFreshAccess {
+                            is_get: *is_get,
+                            target: target_span,
+                            target_type: other_pacing.clone(),
+                            source: span.clone(),
+                            source_type: own_pacing.clone(),
+                        }
+                        .into(),
+                    )
+                } else {
+                    None
+                }
+            })
+            .collect()
     }
 
     fn post_process(
@@ -800,7 +902,7 @@ where
             let close_type = &exp_tt[&keys.close].variant;
 
             let spawn_pacing =
-                (output.spawn().is_none() && spawn_pacing != ConcretePacingType::Constant).then(|| spawn_pacing);
+                (output.spawn().is_none() && spawn_pacing != ConcretePacingType::Constant).then_some(spawn_pacing);
             let spawn_cond = (output.spawn_cond().is_none() && spawn_cond != &positive_top)
                 .then(|| spawn_cond.construct(&[]).expect("variant to not be any"));
             let filter = (output.eval_cond().is_none() && filter_type != &positive_top)
@@ -912,6 +1014,46 @@ where
             }
         }
 
+        // Check that a get / fresh access only occurs between periodic or event-based streams
+        for output in &hir.outputs {
+            // Spawn clause
+            if let Some(spawn) = &output.spawn {
+                let own_pacing: &ConcretePacingType = &pacing_tt[&nid_key[&NodeId::SRef(output.sr)].spawn.0];
+                errors.append(&mut Self::check_get_and_fresh_access(
+                    hir,
+                    pacing_tt,
+                    nid_key,
+                    spawn.expression,
+                    spawn.condition,
+                    own_pacing,
+                ))
+            }
+
+            // Eval clause
+            let eval = &output.eval;
+            let own_pacing: &ConcretePacingType = &pacing_tt[&nid_key[&NodeId::SRef(output.sr)].exp_pacing];
+            errors.append(&mut Self::check_get_and_fresh_access(
+                hir,
+                pacing_tt,
+                nid_key,
+                Some(eval.expr),
+                eval.condition,
+                own_pacing,
+            ));
+
+            // Close clause
+            if let Some(close) = &output.close {
+                let own_pacing: &ConcretePacingType = &pacing_tt[&nid_key[&NodeId::Expr(close.condition)].exp_pacing];
+                errors.append(&mut Self::check_get_and_fresh_access(
+                    hir,
+                    pacing_tt,
+                    nid_key,
+                    None,
+                    Some(close.condition),
+                    own_pacing,
+                ));
+            }
+        }
         errors
     }
 
@@ -1045,11 +1187,15 @@ mod tests {
     }
 
     fn num_errors(spec: &str) -> usize {
+        //let handler = Handler::from(ParserConfig::for_string(spec.to_string()));
         let (spec, _) = setup_ast(spec);
         let mut ltc = LolaTypeChecker::new(&spec);
         match ltc.pacing_type_infer() {
             Ok(_) => 0,
-            Err(e) => e.num_errors(),
+            Err(e) => {
+                //handler.emit_error(&e);
+                e.num_errors()
+            },
         }
     }
 
@@ -2496,6 +2642,62 @@ mod tests {
     #[test]
     fn test_missing_eval_with_apt() {
         let spec = "input a: Int8\noutput b eval @1Hz";
+        assert_eq!(0, num_errors(spec));
+    }
+
+    #[test]
+    fn test_invalid_get() {
+        let spec = "input a: Int8\n\
+        output b \n\
+            eval @1Hz when a.get().defaults(to:42) == 42 with a.get(or: 0)\n\
+        output c(p) \n\
+            spawn @1Hz with a.get(or: 42) when a.get().defaults(to: 2) == 2\n\
+            eval @1Hz with 5\n\
+        output d \n\
+            eval @1Hz with 5\n\
+            close @0.5Hz when a.get(or: 5) == 42";
+        assert_eq!(5, num_errors(spec));
+    }
+
+    #[test]
+    fn test_valid_get() {
+        let spec = "input a: Int8\n\
+        output b \n\
+            eval @a when a.get().defaults(to:42) == 42 with a.get(or: 0)\n\
+        output c(p) \n\
+            spawn @a with a.get(or: 42) when a.get().defaults(to: 2) == 2\n\
+            eval @1Hz with 5\n\
+        output d \n\
+            eval @1Hz with 5\n\
+            close @a when a.get(or: 5) == 42";
+        assert_eq!(0, num_errors(spec));
+    }
+
+    #[test]
+    fn test_invalid_is_fresh() {
+        let spec = "input a: Int8\n\
+        output b \n\
+            eval @1Hz when a.is_fresh() with a.is_fresh()\n\
+        output c(p) \n\
+            spawn @1Hz with a.is_fresh() when a.is_fresh()\n\
+            eval @1Hz with 5\n\
+        output d \n\
+            eval @1Hz with 5\n\
+            close @0.5Hz when a.is_fresh()";
+        assert_eq!(5, num_errors(spec));
+    }
+
+    #[test]
+    fn test_valid_is_fresh() {
+        let spec = "input a: Int8\n\
+        output b \n\
+            eval @a when a.is_fresh() with a.is_fresh()\n\
+        output c(p) \n\
+            spawn @a with a.is_fresh() when a.is_fresh()\n\
+            eval @1Hz with 5\n\
+        output d \n\
+            eval @1Hz with 5\n\
+            close @a when a.is_fresh()";
         assert_eq!(0, num_errors(spec));
     }
 }