+++ crates/rtlola-parser/src/ast.rs
@@ -432,8 +432,10 @@ pub enum StreamAccessKind {
     Sync,
     /// Hold access for *incompatible* stream types, returns previous known value
     Hold,
-    /// Optional access, returns value if it exists
-    Optional,
+    /// Optional access, returns value if it exists, called by `.get()`
+    Get,
+    /// Boolean Typed access, returning true if the target stream received a new value at the current timestamp. Called with `.is_fresh()`.
+    Fresh,
 }
 
 /// Describes the operation used to access a stream with a offset