+++ crates/rtlola-parser/src/parse.rs
@@ -901,7 +901,20 @@ impl RtLolaParser {
                                     }
                                     "get()" => {
                                         assert_eq!(args.len(), 0);
-                                        ExpressionKind::StreamAccess(inner, StreamAccessKind::Optional)
+                                        ExpressionKind::StreamAccess(inner, StreamAccessKind::Get)
+                                    }
+                                    "get(or:)" => {
+                                        assert_eq!(args.len(), 1);
+                                        let lhs = Expression::new(
+                                            self.spec.next_id(),
+                                            ExpressionKind::StreamAccess(inner, StreamAccessKind::Get),
+                                            span.clone(),
+                                        );
+                                        ExpressionKind::Default(Box::new(lhs), Box::new(args[0].clone()))
+                                    }
+                                    "is_fresh()" => {
+                                        assert_eq!(args.len(), 0);
+                                        ExpressionKind::StreamAccess(inner, StreamAccessKind::Fresh)
                                     }
                                     "aggregate(over_discrete:using:)" | "aggregate(over_exactly_discrete:using:)" |"aggregate(over:using:)" | "aggregate(over_exactly:using:)" => {
                                         assert_eq!(args.len(), 2);
@@ -1527,6 +1540,20 @@ mod tests {
     #[test]
     fn build_lookup_expression_hold() {
         let spec = "output s: Int eval with s.offset(by: -1).hold().defaults(to: 3 * 4)\n";
+        let ast = parse(spec);
+        cmp_ast_spec(&ast, spec);
+    }
+
+    #[test]
+    fn build_get_expression_def() {
+        let spec = "output s: Int eval with s.get().defaults(to: -1)\n";
+        let ast = parse(spec);
+        cmp_ast_spec(&ast, spec);
+    }
+
+    #[test]
+    fn build_check_expression_def() {
+        let spec = "input i: Int\noutput s: Int eval @1Hz with if i.is_fresh() then -1 else 1\n";
         let ast = parse(spec);
         cmp_ast_spec(&ast, spec);
     }